requirements.txt
opencv-python 
numpy 
pyyaml 
tqdm 
albumentations